[MIG] partner_multi_relation: Migration to 11.0

+++ partner_multi_relation/models/res_partner_relation_type.py
@@ -1,10 +1,9 @@
-# -*- coding: utf-8 -*-
 # Copyright 2013-2018 Therp BV <https://therp.nl>.
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 """Define the type of relations that can exist between partners."""
-from openerp import _, api, fields, models
-from openerp.exceptions import ValidationError
-from openerp.osv.expression import AND, OR
+from odoo import _, api, fields, models
+from odoo.exceptions import ValidationError
+from odoo.osv.expression import AND, OR
 
 
 HANDLE_INVALID_ONCHANGE = [
@@ -152,7 +151,8 @@ class ResPartnerRelationType(models.Model):
                     # already ended:
                     cutoff_date = fields.Date.today()
                     for relation in invalid_relations:
-                        if relation.date_start >= cutoff_date:
+                        if (relation.date_start and
+                                relation.date_start >= cutoff_date):
                             relation.unlink()
                         elif (not relation.date_end or
                                 relation.date_end > cutoff_date):
